Fetch Proxy Test Finished
Log component should be developed.

+++ Codes/ProxyFetcherConsole/Services/ProxyFetcherService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net.Mail;
+using System.Text;
 using CsvHelper;
 using Yujt.Common.Helper;
 using yujt.common.Proxies;
@@ -14,20 +15,19 @@ namespace ProxyFetcherConsole.Services
     public class ProxyFetcherService : IProxyFetcherService
     {
         private readonly IEmail mEmail163;
-        private string mLocalProxiesPath = FileHelper.GenFileNameInAssemblyDir("/Proxies/proxyList.csv");
+        private string mLocalProxiesPath = FileHelper.GenFileNameInAssemblyDir("Proxies/ProxyList.csv");
 
         private readonly IProxyFetcher mProxyFetcher = new ProxyFetcher();
         private readonly IList<Proxy> mProxies = new List<Proxy>();
         public ProxyFetcherService()
         {
             mEmail163 = new Email163(AppSetting.UserName, AppSetting.Password);
-            //AsynFetchProxies();
         }
 
         public void FetchProxies()
         {
-            var porxiesFromFile = LoadProxiesFromEmail();
-            ValidateProxyFromFile(porxiesFromFile);
+            var porxiesFromEmail = LoadProxiesFromEmail();
+            ValidateProxyFromEmail(porxiesFromEmail);
             EnrichProxiesFromInternet();
 
             UpdateProxyToEmail();
@@ -35,6 +35,7 @@ namespace ProxyFetcherConsole.Services
         #region Private Methods
         private IEnumerable<Proxy> LoadProxiesFromEmail()
         {
+            #region Fetch Proxy File
             var path = Path.GetTempFileName();
             try
             {
@@ -42,8 +43,14 @@ namespace ProxyFetcherConsole.Services
             }
             catch (Exception)
             {
-                path = mLocalProxiesPath;
+                path = FileHelper.GenFileNameInAssemblyDir(mLocalProxiesPath);
             }
+
+            if (!File.Exists(path) || FileHelper.GetFileSize(path) <= 0)
+            {
+                return null;
+            }
+            #endregion Fetch Proxy File
 
             try
             {
@@ -65,13 +72,13 @@ namespace ProxyFetcherConsole.Services
             }
         }
 
-        private void ValidateProxyFromFile(IEnumerable<Proxy> proxiesFromFile)
+        private void ValidateProxyFromEmail(IEnumerable<Proxy> proxiesFromEmail)
         {
-            if (proxiesFromFile == null)
+            if (proxiesFromEmail == null)
             {
                 return;
             }
-            CheckProxies(proxiesFromFile);
+            CheckProxies(proxiesFromEmail);
         }
         private void EnrichProxiesFromInternet()
         {
@@ -97,16 +104,25 @@ namespace ProxyFetcherConsole.Services
 
         private void UpdateProxyToEmail()
         {
-            var bakupFilePath = mLocalProxiesPath + DateTime.Now.ToString("yyyy-MMMM-dd HH:mm-ss");
-            File.Move(mLocalProxiesPath, bakupFilePath);
-            using (var tw = new StreamWriter(mLocalProxiesPath))
+            if (File.Exists(mLocalProxiesPath))
+            {
+                var bakupFilePath = mLocalProxiesPath +"_" + DateTime.Now.ToString("yyyy-MM-dd HH-mm-ss");
+                File.Move(mLocalProxiesPath, bakupFilePath);
+            }
+
+            using (var tw = new StreamWriter(mLocalProxiesPath, false, Encoding.UTF8))
             {
                 var writer = new CsvWriter(tw);
                 writer.WriteRecords(mProxies);
                 tw.Flush();
             }
 
-            var sendMsg = new MailMessage(AppSetting.Subject, mLocalProxiesPath);
+            var sendMsg = new MailMessage(AppSetting.UserName, AppSetting.UserName)
+            {
+                Subject = AppSetting.Subject,
+                Body = AppSetting.Subject
+            };
+            sendMsg.Attachments.Add(new Attachment(mLocalProxiesPath));
 
             mEmail163.Send(sendMsg);
